Add files via upload
The py files with 'ga' in the title are used to tune the elemental matrix using a genetic algorithm.  The py files with 'e2m...' in the titles are then used to apply the tuned matrix to generate mineral predictions.  In all cases, the files with 'function' in the title are not run, but are imported by the '01' and '02' run files.